# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AppConfig {
  oddsjamApiKey: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Sport {
  key: string;
  group: string;
  title: string;
  description: string;
  active: boolean;
  hasOutrights: boolean;
}

export interface BookmakerOdds {
  bookmaker: string;
  bookmakerTitle: string;
  market: string;
  outcome: string;
  price: number;
  /** @nullable */
  point?: number | null;
  lastUpdate: string;
}

export interface GameOdds {
  id: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  commenceTime: string;
  bookmakerOdds: BookmakerOdds[];
}

export interface ArbLeg {
  bookmaker: string;
  bookmakerTitle: string;
  outcome: string;
  price: number;
  stake: number;
  impliedProbability: number;
}

export interface ArbitrageOpportunity {
  id: string;
  gameId: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  commenceTime: string;
  market: string;
  profitPercent: number;
  totalImpliedProbability: number;
  legs: ArbLeg[];
  detectedAt: string;
}

export type OpportunitiesSummarySportBreakdownItem = {
  sport: string;
  count: number;
  avgProfit: number;
};

export type OpportunitiesSummaryMarketBreakdownItem = {
  market: string;
  count: number;
};

export interface OpportunitiesSummary {
  totalOpportunities: number;
  averageProfitPercent: number;
  bestProfitPercent: number;
  sportBreakdown: OpportunitiesSummarySportBreakdownItem[];
  marketBreakdown: OpportunitiesSummaryMarketBreakdownItem[];
}

export interface Alert {
  id: number;
  /** @nullable */
  sport?: string | null;
  /** @nullable */
  market?: string | null;
  minProfitPercent: number;
  createdAt: string;
}

export interface CreateAlertBody {
  /** @nullable */
  sport?: string | null;
  /** @nullable */
  market?: string | null;
  minProfitPercent: number;
}

export type GetOddsParams = {
  /**
   * Sport key (e.g. americanfootball_nfl)
   */
  sport: string;
  /**
   * Comma-separated market keys (e.g. h2h,spreads,totals)
   */
  markets?: string;
  /**
   * Comma-separated bookmaker keys to filter
   */
  bookmakers?: string;
};

export type GetArbitrageOpportunitiesParams = {
  /**
   * Filter by sport key
   */
  sport?: string;
  /**
   * Minimum profit percentage (default 0)
   */
  minProfit?: number;
  /**
   * Filter by market type (h2h, spreads, totals)
   */
  market?: string;
};