# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns configuration values needed by the frontend (API keys, etc.)
 * @summary Get frontend configuration
 */
export const GetConfigResponse = zod.object({
  oddsjamApiKey: zod.string(),
});

/**
 * Returns list of sports supported by OddsJam
 * @summary Get available sports
 */
export const GetSportsResponseItem = zod.object({
  key: zod.string(),
  group: zod.string(),
  title: zod.string(),
  description: zod.string(),
  active: zod.boolean(),
  hasOutrights: zod.boolean(),
});
export const GetSportsResponse = zod.array(GetSportsResponseItem);

/**
 * Fetches live odds from OddsJam for arbitrage analysis
 * @summary Get live odds for a sport
 */
export const GetOddsQueryParams = zod.object({
  sport: zod.coerce.string().describe("Sport key (e.g. americanfootball_nfl)"),
  markets: zod.coerce
    .string()
    .optional()
    .describe("Comma-separated market keys (e.g. h2h,spreads,totals)"),
  bookmakers: zod.coerce
    .string()
    .optional()
    .describe("Comma-separated bookmaker keys to filter"),
});

export const GetOddsResponseItem = zod.object({
  id: zod.string(),
  sport: zod.string(),
  homeTeam: zod.string(),
  awayTeam: zod.string(),
  commenceTime: zod.string(),
  bookmakerOdds: zod.array(
    zod.object({
      bookmaker: zod.string(),
      bookmakerTitle: zod.string(),
      market: zod.string(),
      outcome: zod.string(),
      price: zod.number(),
      point: zod.number().nullish(),
      lastUpdate: zod.string(),
    }),
  ),
});
export const GetOddsResponse = zod.array(GetOddsResponseItem);

/**
 * Returns computed arbitrage opportunities sorted by profit percentage
 * @summary Get arbitrage opportunities
 */
export const GetArbitrageOpportunitiesQueryParams = zod.object({
  sport: zod.coerce.string().optional().describe("Filter by sport key"),
  minProfit: zod.coerce
    .number()
    .optional()
    .describe("Minimum profit percentage (default 0)"),
  market: zod.coerce
    .string()
    .optional()
    .describe("Filter by market type (h2h, spreads, totals)"),
});

export const GetArbitrageOpportunitiesResponseItem = zod.object({
  id: zod.string(),
  gameId: zod.string(),
  sport: zod.string(),
  homeTeam: zod.string(),
  awayTeam: zod.string(),
  commenceTime: zod.string(),
  market: zod.string(),
  profitPercent: zod.number(),
  totalImpliedProbability: zod.number(),
  legs: zod.array(
    zod.object({
      bookmaker: zod.string(),
      bookmakerTitle: zod.string(),
      outcome: zod.string(),
      price: zod.number(),
      stake: zod.number(),
      impliedProbability: zod.number(),
    }),
  ),
  detectedAt: zod.string(),
});
export const GetArbitrageOpportunitiesResponse = zod.array(
  GetArbitrageOpportunitiesResponseItem,
);

/**
 * Returns aggregate stats on current arbitrage opportunities
 * @summary Get summary stats for arbitrage opportunities
 */
export const GetOpportunitiesSummaryResponse = zod.object({
  totalOpportunities: zod.number(),
  averageProfitPercent: zod.number(),
  bestProfitPercent: zod.number(),
  sportBreakdown: zod.array(
    zod.object({
      sport: zod.string(),
      count: zod.number(),
      avgProfit: zod.number(),
    }),
  ),
  marketBreakdown: zod.array(
    zod.object({
      market: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * Returns user-configured arbitrage alerts
 * @summary Get saved alerts
 */
export const GetAlertsResponseItem = zod.object({
  id: zod.number(),
  sport: zod.string().nullish(),
  market: zod.string().nullish(),
  minProfitPercent: zod.number(),
  createdAt: zod.string(),
});
export const GetAlertsResponse = zod.array(GetAlertsResponseItem);

/**
 * Save an arbitrage alert configuration
 * @summary Create an alert
 */
export const CreateAlertBody = zod.object({
  sport: zod.string().nullish(),
  market: zod.string().nullish(),
  minProfitPercent: zod.number(),
});

/**
 * @summary Delete an alert
 */
export const DeleteAlertParams = zod.object({
  id: zod.coerce.number(),
});