# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  /** @nullable */
  sport?: string | null;
  /** @nullable */
  market?: string | null;
  minProfitPercent: number;
  createdAt: string;
}