# lib/api-zod/src/generated/types/arbitrageOpportunity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */
import type { ArbLeg } from "./arbLeg";

export interface ArbitrageOpportunity {
  id: string;
  gameId: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  commenceTime: string;
  market: string;
  profitPercent: number;
  totalImpliedProbability: number;
  legs: ArbLeg[];
  detectedAt: string;
}