# lib/api-zod/src/generated/types/bookmakerOdds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */

export interface BookmakerOdds {
  bookmaker: string;
  bookmakerTitle: string;
  market: string;
  outcome: string;
  price: number;
  /** @nullable */
  point?: number | null;
  lastUpdate: string;
}