# lib/api-zod/src/generated/types/gameOdds.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */
import type { BookmakerOdds } from "./bookmakerOdds";

export interface GameOdds {
  id: string;
  sport: string;
  homeTeam: string;
  awayTeam: string;
  commenceTime: string;
  bookmakerOdds: BookmakerOdds[];
}