# lib/api-zod/src/generated/types/opportunitiesSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunitiesSummaryMarketBreakdownItem } from "./opportunitiesSummaryMarketBreakdownItem";
import type { OpportunitiesSummarySportBreakdownItem } from "./opportunitiesSummarySportBreakdownItem";

export interface OpportunitiesSummary {
  totalOpportunities: number;
  averageProfitPercent: number;
  bestProfitPercent: number;
  sportBreakdown: OpportunitiesSummarySportBreakdownItem[];
  marketBreakdown: OpportunitiesSummaryMarketBreakdownItem[];
}