# lib/api-zod/src/generated/types/sport.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sports betting arbitrage finder API
 * OpenAPI spec version: 0.1.0
 */

export interface Sport {
  key: string;
  group: string;
  title: string;
  description: string;
  active: boolean;
  hasOutrights: boolean;
}